Moved DownsampleTransform and SuperResolutionDataset classes from dataset.py to modules.py.
